+++ src/libYARP_sig/src/Matrix.cpp
@@ -387,7 +387,7 @@ bool Matrix::setSubcol(const Vector &v, int r, int c)
     if(r<0 || r+v.size()>=(size_t)nrows || c<0 || c>=ncols)
         return false;
 
-    for(size_t i=0;i<v.size();i++)
+    for(int i=0;i<(int)v.size();i++)
         (*this)[r+i][c] = v[i];
     return true;
 }
